refactor(api): broken down the stats.py into files by topic
