+++ api/src/error_handlers.py
@@ -1,11 +1,9 @@
-"""JSON error handlers matching the flask-smorest `Error` envelope.
+"""JSON error handlers matching the flask-smorest Error envelope.
 
-* Return the same JSON shape on every layer (`{code, status, message, errors}`).
-* Log 5xx with `exc_info` plus request method + path.
-* Log 4xx at INFO with method + path so we can see scanning / bad clients.
+Every failure, whatever raised it, comes back as
+`{code, status, message, errors}` so clients parse one shape.
 
-The client IP (`request.remote_addr`) is intentionally omitted from every log
-line here.
+Log lines never include request.remote_addr: attacker IPs stay out of the logs.
 """
 
 from __future__ import annotations
@@ -31,11 +29,10 @@ def init_error_handlers(app: Flask) -> None:
         code = exc.code or 500
         status = exc.name or "Error"
         message = exc.description or status
+        # flask-smorest passes marshmallow validation output through exc.data;
+        # keep its errors/message rather than the generic HTTP description.
         body = getattr(exc, "data", None) or {}
-        # flask-smorest stuffs marshmallow validation messages under
-        # `exc.data['errors']`; preserve that shape.
         errors = body.get("errors") if isinstance(body, dict) else None
-        # smorest also passes its own `message` field via `exc.data`.
         if isinstance(body, dict) and isinstance(body.get("message"), str):
             message = body["message"]
