+++ api/src/schemas/common.py
@@ -23,11 +23,10 @@ class CountryCodeField(fields.String):
 
 
 def country_filter_field() -> fields.Str:
-    """Fresh optional alpha-2 country filter field (one instance per schema).
+    """Optional alpha-2 country filter.
 
-    Marshmallow fields are bound to their owning schema, so each query schema
-    needs its own instance; this factory keeps the regex and metadata defined
-    in exactly one place.
+    A factory, not a shared instance: marshmallow binds a field to its owning
+    schema, so reusing one across schemas breaks.
     """
     return CountryCodeField(
         load_default=None,
@@ -43,7 +42,7 @@ def country_filter_field() -> fields.Str:
 def top_n_field(
     default: int, description: str = "Number of top entries to return"
 ) -> fields.Int:
-    """Fresh top-N limit field, see :func:`country_filter_field` for why fresh."""
+    """Top-N limit field; a factory for the same reason as above."""
     return fields.Int(
         load_default=default,
         validate=validate.Range(min=1, max=100),
@@ -52,11 +51,10 @@ def top_n_field(
 
 
 def country_or_unknown_field() -> fields.Str:
-    """Country filter that also accepts `"??"` -- the geo-less bucket.
+    """Country filter that also accepts "??".
 
-    Same as :func:`country_filter_field`, but the credential/ASN leaderboards
-    let the Countries page drill into the "Unknown" row (sessions whose source
-    IP has no resolved country), addressed by the `"??"` sentinel.
+    The credential and ASN leaderboards are reachable by clicking a row on the
+    Countries page, including its Unknown row - "??" addresses that bucket.
     """
     return CountryCodeField(
         load_default=None,
@@ -75,8 +73,8 @@ def country_or_unknown_field() -> fields.Str:
 class BaseSchema(Schema):
     """Shared base for every schema.
 
-    Marshmallow 4 preserves field declaration order natively, so OpenAPI
-    properties stay stable across regenerations with no `Meta` config needed.
+    Empty on purpose: marshmallow 4 keeps field declaration order, so the
+    generated OpenAPI is stable without any Meta config.
     """
