+++ api/src/security_headers.py
@@ -1,6 +1,7 @@
 """App-layer security headers.
 
-Proxy already sets these in production. Keeping it for local debug.
+nginx sets these in production; these are the fallback for running the app
+directly. setdefault everywhere, so the proxy's values always win.
 """
 
 from __future__ import annotations
@@ -9,7 +10,7 @@ from flask import Flask, Response, request
 
 
 def init_security_headers(app: Flask) -> None:
-    # CSP and HSTS owned by nginx .
+    # CSP and HSTS are nginx-only: both need the deployed origin to be right.
     @app.after_request
     def _set_headers(response: Response) -> Response:  # pyright: ignore[reportUnusedFunction]
         response.headers.setdefault("X-Content-Type-Options", "nosniff")
@@ -18,7 +19,7 @@ def init_security_headers(app: Flask) -> None:
             "Referrer-Policy", "strict-origin-when-cross-origin"
         )
         if response.mimetype == "application/json":
-            # The OpenAPI spec and error responses must not be cached.
+            # Stats tolerate 30s of staleness; a cached spec or error does not.
             if request.path.endswith("/openapi.json") or response.status_code >= 400:
                 response.headers.setdefault("Cache-Control", "no-store")
             else:
